+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `to_tex` for gridded perms.
+- `to_tikz` for gridded perms.
+- `to_svg` for gridded perms.
 
 ## [2.3.1] - 2020-09-11
 ### Fixed